add: test for v0.2
